+++ lib/sequent/test.rb
@@ -3,4 +3,4 @@
 require_relative 'sequent'
 require_relative 'test/command_handler_helpers'
 require_relative 'test/event_stream_helpers'
-require_relative 'test/event_handler_helpers'
+require_relative 'test/workflow_helpers'